fix: allow copy of iter wrapper if wraps rusty iter
